# Faster tokenizer lookahead