--- a/packages/babel-parser/test/fixtures/typescript/export/internal-comments/output.json
+++ b/packages/babel-parser/test/fixtures/typescript/export/internal-comments/output.json
@@ -1,0 +1,242 @@
+{
+  "type": "File",
+  "start":0,"end":137,"loc":{"start":{"line":1,"column":0},"end":{"line":2,"column":43}},
+  "program": {
+    "type": "Program",
+    "start":0,"end":137,"loc":{"start":{"line":1,"column":0},"end":{"line":2,"column":43}},
+    "sourceType": "module",
+    "interpreter": null,
+    "body": [
+      {
+        "type": "ExportNamedDeclaration",
+        "start":6,"end":93,"loc":{"start":{"line":1,"column":6},"end":{"line":1,"column":93}},
+        "leadingComments": [
+          {
+            "type": "CommentBlock",
+            "value": "1",
+            "start":0,"end":5,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":5}}
+          }
+        ],
+        "trailingComments": [
+          {
+            "type": "CommentBlock",
+            "value": "1",
+            "start":94,"end":99,"loc":{"start":{"line":2,"column":0},"end":{"line":2,"column":5}}
+          }
+        ],
+        "exportKind": "value",
+        "specifiers": [
+          {
+            "type": "ExportSpecifier",
+            "start":27,"end":28,"loc":{"start":{"line":1,"column":27},"end":{"line":1,"column":28}},
+            "leadingComments": [
+              {
+                "type": "CommentBlock",
+                "value": "2",
+                "start":13,"end":18,"loc":{"start":{"line":1,"column":13},"end":{"line":1,"column":18}}
+              },
+              {
+                "type": "CommentBlock",
+                "value": "3",
+                "start":21,"end":26,"loc":{"start":{"line":1,"column":21},"end":{"line":1,"column":26}}
+              }
+            ],
+            "trailingComments": [
+              {
+                "type": "CommentBlock",
+                "value": "4",
+                "start":29,"end":34,"loc":{"start":{"line":1,"column":29},"end":{"line":1,"column":34}}
+              },
+              {
+                "type": "CommentBlock",
+                "value": "5",
+                "start":36,"end":41,"loc":{"start":{"line":1,"column":36},"end":{"line":1,"column":41}}
+              }
+            ],
+            "local": {
+              "type": "Identifier",
+              "start":27,"end":28,"loc":{"start":{"line":1,"column":27},"end":{"line":1,"column":28},"identifierName":"A"},
+              "name": "A"
+            },
+            "exported": {
+              "type": "Identifier",
+              "start":27,"end":28,"loc":{"start":{"line":1,"column":27},"end":{"line":1,"column":28},"identifierName":"A"},
+              "name": "A"
+            }
+          },
+          {
+            "type": "ExportSpecifier",
+            "start":42,"end":60,"loc":{"start":{"line":1,"column":42},"end":{"line":1,"column":60}},
+            "trailingComments": [
+              {
+                "type": "CommentBlock",
+                "value": "8",
+                "start":61,"end":66,"loc":{"start":{"line":1,"column":61},"end":{"line":1,"column":66}}
+              },
+              {
+                "type": "CommentBlock",
+                "value": "9",
+                "start":69,"end":74,"loc":{"start":{"line":1,"column":69},"end":{"line":1,"column":74}}
+              },
+              {
+                "type": "CommentBlock",
+                "value": "10",
+                "start":80,"end":86,"loc":{"start":{"line":1,"column":80},"end":{"line":1,"column":86}}
+              }
+            ],
+            "local": {
+              "type": "Identifier",
+              "start":42,"end":43,"loc":{"start":{"line":1,"column":42},"end":{"line":1,"column":43},"identifierName":"B"},
+              "trailingComments": [
+                {
+                  "type": "CommentBlock",
+                  "value": "6",
+                  "start":44,"end":49,"loc":{"start":{"line":1,"column":44},"end":{"line":1,"column":49}}
+                }
+              ],
+              "name": "B"
+            },
+            "exported": {
+              "type": "Identifier",
+              "start":59,"end":60,"loc":{"start":{"line":1,"column":59},"end":{"line":1,"column":60},"identifierName":"C"},
+              "leadingComments": [
+                {
+                  "type": "CommentBlock",
+                  "value": "6",
+                  "start":44,"end":49,"loc":{"start":{"line":1,"column":44},"end":{"line":1,"column":49}}
+                },
+                {
+                  "type": "CommentBlock",
+                  "value": "7",
+                  "start":53,"end":58,"loc":{"start":{"line":1,"column":53},"end":{"line":1,"column":58}}
+                }
+              ],
+              "name": "C"
+            }
+          }
+        ],
+        "source": {
+          "type": "StringLiteral",
+          "start":87,"end":92,"loc":{"start":{"line":1,"column":87},"end":{"line":1,"column":92}},
+          "extra": {
+            "rawValue": "foo",
+            "raw": "\"foo\""
+          },
+          "value": "foo"
+        },
+        "declaration": null
+      },
+      {
+        "type": "ExportAllDeclaration",
+        "start":100,"end":137,"loc":{"start":{"line":2,"column":6},"end":{"line":2,"column":43}},
+        "leadingComments": [
+          {
+            "type": "CommentBlock",
+            "value": "1",
+            "start":94,"end":99,"loc":{"start":{"line":2,"column":0},"end":{"line":2,"column":5}}
+          }
+        ],
+        "exportKind": "value",
+        "source": {
+          "type": "StringLiteral",
+          "start":132,"end":137,"loc":{"start":{"line":2,"column":38},"end":{"line":2,"column":43}},
+          "leadingComments": [
+            {
+              "type": "CommentBlock",
+              "value": "2",
+              "start":107,"end":112,"loc":{"start":{"line":2,"column":13},"end":{"line":2,"column":18}}
+            },
+            {
+              "type": "CommentBlock",
+              "value": "3",
+              "start":115,"end":120,"loc":{"start":{"line":2,"column":21},"end":{"line":2,"column":26}}
+            },
+            {
+              "type": "CommentBlock",
+              "value": "4",
+              "start":126,"end":131,"loc":{"start":{"line":2,"column":32},"end":{"line":2,"column":37}}
+            }
+          ],
+          "extra": {
+            "rawValue": "foo",
+            "raw": "\"foo\""
+          },
+          "value": "foo"
+        }
+      }
+    ],
+    "directives": []
+  },
+  "comments": [
+    {
+      "type": "CommentBlock",
+      "value": "1",
+      "start":0,"end":5,"loc":{"start":{"line":1,"column":0},"end":{"line":1,"column":5}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "2",
+      "start":13,"end":18,"loc":{"start":{"line":1,"column":13},"end":{"line":1,"column":18}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "3",
+      "start":21,"end":26,"loc":{"start":{"line":1,"column":21},"end":{"line":1,"column":26}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "4",
+      "start":29,"end":34,"loc":{"start":{"line":1,"column":29},"end":{"line":1,"column":34}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "5",
+      "start":36,"end":41,"loc":{"start":{"line":1,"column":36},"end":{"line":1,"column":41}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "6",
+      "start":44,"end":49,"loc":{"start":{"line":1,"column":44},"end":{"line":1,"column":49}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "7",
+      "start":53,"end":58,"loc":{"start":{"line":1,"column":53},"end":{"line":1,"column":58}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "8",
+      "start":61,"end":66,"loc":{"start":{"line":1,"column":61},"end":{"line":1,"column":66}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "9",
+      "start":69,"end":74,"loc":{"start":{"line":1,"column":69},"end":{"line":1,"column":74}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "10",
+      "start":80,"end":86,"loc":{"start":{"line":1,"column":80},"end":{"line":1,"column":86}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "1",
+      "start":94,"end":99,"loc":{"start":{"line":2,"column":0},"end":{"line":2,"column":5}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "2",
+      "start":107,"end":112,"loc":{"start":{"line":2,"column":13},"end":{"line":2,"column":18}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "3",
+      "start":115,"end":120,"loc":{"start":{"line":2,"column":21},"end":{"line":2,"column":26}}
+    },
+    {
+      "type": "CommentBlock",
+      "value": "4",
+      "start":126,"end":131,"loc":{"start":{"line":2,"column":32},"end":{"line":2,"column":37}}
+    }
+  ]
+}