… packages/babel-parser/test/fixtures/typescript/export/internal-comments/input.ts
… packages/babel-parser/test/fixtures/typescript/import/internal-comments/input.ts
… packages/babel-parser/test/fixtures/typescript/import/internal-comments/output.json