<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
Currently we clone the whole parser state when doing `lookahead()`, this is expensive because the `State` is large and in the past we have replaced some `lookahead()` usage by `lookaheadCharCode()` (#10371). 

In this PR we limit the returned structure of `tokenizer#lookahead()`, based on the observations that 1) in most cases we retrieve only `type` and `value` from the result of `lookahead()`, 2) the tokenizer only depends on a subset of parser state.

We use the Flow arrow function types `type A = (x) => void` as a benchmark case because parsing such types invokes two `lookahead()` calls. As we can see parsing such types are 200% faster than the baseline. 

```
$ node --predictable ./many-arrow-function-flow-types/bench.mjs
baseline 256 arrow function types: 506 ops/sec ±1.06% (1.976ms)
baseline 512 arrow function types: 256 ops/sec ±0.72% (3.906ms)
baseline 1024 arrow function types: 121 ops/sec ±0.63% (8.266ms)
baseline 2048 arrow function types: 50.41 ops/sec ±1.4% (20ms)
current 256 arrow function types: 1537 ops/sec ±1.25% (0.651ms)
current 512 arrow function types: 733 ops/sec ±0.77% (1.364ms)
current 1024 arrow function types: 359 ops/sec ±0.3% (2.786ms)
current 2048 arrow function types: 171 ops/sec ±1% (5.859ms)
```

Surprisingly, if we replace `lookahead()` (shown as "baseline") by `lookaheadCharCode()` (shown as "current"), performance is almost halved. It means [`skipWhiteSpace` regex](https://github.com/babel/babel/blob/12190042e6d62d10272a50bdc624e0db327a705a/packages/babel-parser/src/util/whitespace.js#L24) may be slower than our handwritten [`skipSpace`](https://github.com/babel/babel/blob/12190042e6d62d10272a50bdc624e0db327a705a/packages/babel-parser/src/tokenizer/index.js#L345) method when there aren't many whitespaces to skip.
```
baseline 256 arrow function types: 1533 ops/sec ±1.5% (0.652ms)
baseline 512 arrow function types: 721 ops/sec ±1.3% (1.386ms)
baseline 1024 arrow function types: 353 ops/sec ±1.21% (2.833ms)
baseline 2048 arrow function types: 179 ops/sec ±0.9% (5.588ms)
current 256 arrow function types: 861 ops/sec ±3.14% (1.162ms)
current 512 arrow function types: 445 ops/sec ±2.13% (2.245ms)
current 1024 arrow function types: 218 ops/sec ±2.33% (4.584ms)
current 2048 arrow function types: 101 ops/sec ±2.61% (9.877ms)
``` 

<a href="https://gitpod.io/#https://github.com/babel/babel/pull/13341"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>