@nicolo-ribaudo 2021-05-21
Can we also add a non-test262 test for [`6349459` (#13341)](https://github.com/babel/babel/pull/13341/commits/6349459955082ab355515b48057619b9796442bf)? Even if we have the test262 test, I'd like our test suite to cover every possible code path.