@JLHwung 2021-05-26
@KFlash Nope, we have only latest `@babel/standlone` build, so currently will have to clone the repo and build it by yourself. However I don't expect it will improve that much on average cases -- the `lookahead()` is only used in Flow and TypeScript plugin, which also used `state.clone` in `tryParse` heavily.

> Here you got 2x state.end. It's better to allocate some memory for a var and avoid duplicate code and 2x. object access on each run

You are right, V8 generates two object access here.
```asm
                  -- <babel/packages/babel-parser/lib/index.js:8479:18> inlined at <babel/packages/babel-parser/lib/index.js:8489:23> --
0x1fafde11f291   111  4d8bb90f010000 REX.W movq r15,[r9+0x10f]
0x1fafde11f298   118  4d897b37       REX.W movq [r11+0x37],r15
                  -- <babel/packages/babel-parser/lib/index.js:8480:25> inlined at <babel/packages/babel-parser/lib/index.js:8489:23> --
0x1fafde11f29c   11c  4d8bb90f010000 REX.W movq r15,[r9+0x10f]
0x1fafde11f2a3   123  4d897b3f       REX.W movq [r11+0x3f],r15
```

For those member expression in object literal, e.g. 
```js
{
  pos: state.pos
}
```

V8 generates extra initialization machine code, which could be eliminated when the object literal contains only literals / identifiers.
```asm
0x1fafde11f238    b8  49c7432f00000000 REX.W movq [r11+0x2f],0x0
```