@KFlash 2021-05-26
Regarding the flow and typescript. They are always doing a call to their parent class - witch in turn kills perf. The TS plugin is the worse candidate if you compare them.

why do you need that extra 'state.pos'?  Why not flatten it?

Did you check with V8 all the extra expenses you get with extending a class? 

The ideal solution is to skip string comparison and let everything become a series of numbers. See my Kataw parser as an example.